requirements.txt
langchain
python-dotenv
beautifulsoup4
playwright
html2text
boto3
streamlit
